move delete_unlinked_assets
